# Optimize the hasMatchingTag function in memory.go