Using regexp to match the exact value in `itemTags` hurts performance. In addition, it's unnecessary because `==` does the same thing.

I added a test to verify that this change does chage the logic of `hasMatchingTag`.

To prove that mine implementation is running faster I called it `hasMatchinTag1` and ran benchmarks:
```Go
func generateTags(num int) map[string]string {
        m := map[string]string{}
        for i := 0; i < num; i++ {
                m[fmt.Sprintf("k%d", i)] = "value"
        }
        return m
}

func BenchmarkHasMatchingTag_10(b *testing.B) {
        tags := generateTags(10)
        itemTags := generateTags(10)

        for i := 0; i < b.N; i++ {
                hasMatchingTag(tags, itemTags)
        }
}

func BenchmarkHasMatchingTag1_10(b *testing.B) {
        tags := generateTags(10)
        itemTags := generateTags(10)

        for i := 0; i < b.N; i++ {
                hasMatchingTag1(tags, itemTags)
        }
}

func BenchmarkHasMatchingTag_100(b *testing.B) {
        tags := generateTags(100)
        itemTags := generateTags(100)

        for i := 0; i < b.N; i++ {
                hasMatchingTag(tags, itemTags)
        }
}

func BenchmarkHasMatchingTag1_100(b *testing.B) {
        tags := generateTags(100)
        itemTags := generateTags(100)

        for i := 0; i < b.N; i++ {
                hasMatchingTag1(tags, itemTags)
        }
}
...
```

the results were:
```shell
BenchmarkHasMatchingTag_10-4               20000             69079 ns/op           57280 B/op        580 allocs/op
BenchmarkHasMatchingTag1_10-4            3000000               475 ns/op               0 B/op          0 allocs/op
BenchmarkHasMatchingTag_100-4               2000            701758 ns/op          572813 B/op       5800 allocs/op
BenchmarkHasMatchingTag1_100-4            300000              5002 ns/op               0 B/op          0 allocs/op
BenchmarkHasMatchingTag_1000-4               200           7158635 ns/op         5729856 B/op      58020 allocs/op
BenchmarkHasMatchingTag1_1000-4            20000             68161 ns/op              18 B/op          0 allocs/op
BenchmarkHasMatchingTag_10000-4               20          96200967 ns/op        57424807 B/op     582057 allocs/op
BenchmarkHasMatchingTag1_10000-4            2000            894158 ns/op            1445 B/op         20 allocs/op
```

Signed-off-by: Boaz Shuster <ripcurld.github@gmail.com>