@yaacov 2017-12-04
we need to support regexp tags ...
e.g. ".*com" to match "example.com" and "hello|world" to match "hello" or "world"

we need to think abou it more ...